ponderar consertado

+++ SRA/src/model/IRede.java
@@ -29,5 +29,5 @@ public interface IRede {
     
     int ponderar(Pessoa p1, Pessoa p2);
     
-    int gerarPontuacao(Pessoa p1, Pessoa p2);
+    double gerarPontuacao(Pessoa p1, Pessoa p2);
 }
